requirements.txt
absl-py==2.1.0
alembic==1.13.3
altair==5.4.1
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
appdirs==1.4.4
astunparse==1.6.3
attrs==24.2.0
autograd==1.7.0
autoray==0.6.12
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
dill==0.3.8
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.0.3
Flask-Dropzone==2.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fpdf==1.7.2
fsspec==2024.9.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
libclang==18.1.1
looseversion==1.3.0
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgspec==0.18.6
namex==0.0.8
narwhals==1.8.3
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
PennyLane==0.38.0
PennyLane_Lightning==0.38.0
pillow==10.4.0
portpicker==1.6.0
protobuf==4.25.5
prov==2.0.1
psutil==6.0.0
pydot==3.0.1
Pygments==2.18.0
pyparsing==3.1.4
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rustworkx==0.15.1
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traits==6.3.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vl-convert-python==1.6.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
